requirements.txt
aiohappyeyeballs==2.3.5
aiohttp==3.10.3
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
arabic-reshaper==3.0.0
arxiv==2.1.3
asgiref==3.8.1
asn1crypto==1.5.1
asttokens==2.4.1
attrs==24.2.0
-e git+ssh://git@github.com/mhmaguire/linkedIn_auto_jobs_applier_with_AI.git@2b21193a1f0ceef5665dcf49d0ab942961df1f68#egg=auto_resume
bidict==0.23.1
blinker==1.8.2
certifi==2024.7.4
cffi==1.17.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.0
cssselect2==0.7.0
dataclasses-json==0.6.7
decorator==5.1.1
distro==1.9.0
executing==2.0.1
feedparser==6.0.11
Flask==3.0.3
Flask-SocketIO==5.3.6
frozenlist==1.4.1
greenlet==3.0.3
grpcio==1.66.0
grpcio-tools==1.66.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
html5lib==1.1
httpcore==1.0.5
httpx==0.27.0
hyperframe==6.0.1
idna==3.7
ipython==8.26.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.2.11
langchain-community==0.2.10
langchain-core==0.2.34
langchain-experimental==0.0.64
langchain-openai==0.1.17
langchain-qdrant==0.1.3
langchain-text-splitters==0.2.2
langgraph==0.1.19
langsmith==0.1.93
Levenshtein==0.25.1
lxml==5.3.0
Markdown==3.7
MarkupSafe==2.1.5
marshmallow==3.21.3
matplotlib-inline==0.1.7
multidict==6.0.5
mypy-extensions==1.0.0
neo4j==5.23.1
nodeenv==1.9.1
numpy==1.26.4
openai==1.37.1
orjson==3.10.7
oscrypto==1.3.0
outcome==1.3.0.post0
packaging==24.1
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
portalocker==2.10.1
prisma==0.14.0
prompt_toolkit==3.0.47
protobuf==5.27.3
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
pyHanko==0.25.1
pyhanko-certvalidator==0.26.3
PyMuPDF==1.24.9
PyMuPDFb==1.24.9
pypdf==4.3.1
pypng==0.20220715.0
PySocks==1.7.1
python-bidi==0.6.0
python-dotenv==1.0.1
python-engineio==4.9.1
python-socketio==5.11.3
pytz==2024.1
PyYAML==6.0.2
pyzmq==26.1.0
qdrant-client==1.11.0
qrcode==7.4.2
rapidfuzz==3.9.6
regex==2024.7.24
reportlab==4.2.2
requests==2.32.3
selenium==4.9.1
sgmllib3k==1.0.0
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.32
stack-data==0.6.3
svglib==1.5.1
tenacity==8.5.0
tiktoken==0.7.0
tinycss2==1.3.0
tomlkit==0.13.0
tqdm==4.66.5
traitlets==5.14.3
trio==0.26.2
trio-websocket==0.11.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzlocal==5.2
uritools==4.0.3
urllib3==2.2.2
wcwidth==0.2.13
webdriver-manager==4.0.2
webencodings==0.5.1
Werkzeug==3.0.3
wsproto==1.2.0
xhtml2pdf==0.2.16
yarl==1.9.4
